SNAP-284: Ameliorated the favicon, especially, we needed to support the .ico extension / file format.
Changed the implementation to use the /favicon.ico as an attachment instead of a special page.
Added support for .ico/x-icon images in the editor.
Added debug and warning messages in the function handling images in the editor.
Fixed the path in case the ipath is the root page, it would double the starting slash ("//favicon.ico") which does not work since // in HTML is now viewed as domain name and not root path.
Fixed the parse_out_inline_img() function so we properly detect changes, because it was not always increment the "changed" variable as expected.
Applied a fix to the parse_out_inline_img() in case  the only element was the img tag because then it was becoming an empty element tag.
Fixed the owner of the /favicon.ico and /default-favicon.ico pages back to "favicon" as it ought to be (we may fix that later with proper attachments instead...)
Tried to add more support of the magic file language, but that fails, it is actually way more complicated than I first thought. Instead I added a snap-basics file with whatever format we want to support and is not available in a "compatible" file.
Many clean ups and small enhancements and bumped copyright date to 2016 in all modified files that include a notice.