SNAP-300: Started adding the time tracker plugin.
Fixed the backend_nice variable definition, it has to include the fullname (i.e. "list::pagelist" and not just "pagelist".)
Removed sendmail_udp_signal from the snapserver.conf, I still need to check on images_udp_signal and snapserver_udp_signal.
Prevent the error output whenever a backend action is not defined in a plugin.
Fixed the "last_day_of_month()" function for year 1752, it was considered to have 29 days in February.
Fixed the initialization of layouts from resources. It would not occur properly because the theme was never added under /admin/layouts.
Broken up the list generation from the list plugin in a separate function which we can call from other places.
Various clean ups.