SNAP-300: Added content types for the calendar and day pages in the timetracker.
Implemented the "Add Self" button so an administrator can add himself to the timetracker system.
Implemented the next/previous month/year buttons with the full update of the calendar inline.
Made the calendar look decent.
Fixed the calendar so the correct dates get shown (cpp, xsl).
Fixed the retrieval of the date to show in the calendar.
Fixed the adding of day_tag so they get added to the line_tag instead of days_tag.
Properly process the calendar data through the calendar-parser.xsl.
Various cleanups.