SNAP-300: Added the timetracker plugin.
SNAP-301: Added the bookkeeping plugin so we can have clients defined for the timetracker.
EX-1: Added a logo to the Char Chart plugin.
Fixed the bare theme so the style.css is added as it should and the editor is not forced anymore (since that bug was fixed in core.)
SNAP-227: dropdown default value is now checked against text() instead of node() so it works better in more cases.
Removed the required asterisk from the dropdown. At this time we are not dealing with labels so it does not get placed properly.
Moved the beforeHide() function from finball as the default in the editor.
Moved the CSS for the Editor confirm closure of a popup so it looks acceptable.
Applied a fix so dictionary corrections (right click + select proper spelling) gets detected and the Save button appears as expected.
Fixed the JavaScript openDropdown() function so it finds the iframe without using "finball" as the name since this code is not specific to finball!
Got the user-messages box to appear with a size that fits when opened in a popup.
Applied a fix to the Popup window used to display a messageBox in order to give each button its name in the href after the '#' character.
Now I put the id of the iframe as the name as well so that way we can access it that way too.
SNAP-289: applied a few changes to the magic-to-js code as underlined by Coverity.
Various clean ups.