EX-8: SNAP-220: Made the minimum and maximum length work against all text fields.
Implemented the min/max length against the stripped (minlength) and non-stripped (absoluteminlength) value.
Implemented the erroneous class on widgets that are detected as having a problem either right here or from the server.
Implemented a getLabel() on EditorWidget so errors can reference the label and not the technical name of the widget.
Finally fixed the default value of a dropdown which is used as a "Please select something" message so it does not get returned anymore.
Added necessary code to automatically hide the message box on a successful save in the editor.
Fixed the serverAccessComplete() in server-access.js, it would eventually display an error message even though it would be redirecting right after that.
Implemented a fix in the output.js to display widget labels instead of their technical names.

SNAP-291: Added icons for the header and images plugins.
Fixed the locale_settings dependencies, it does depend on locale_widgets.
Added a line-edit for the locale so we can enter something because "" fails in many ways!

Various cleanups.