SNAP-286: Attachments auto-generated by backends now get deleted automatically as they ought to be.
Fixed the test to know whether there is an MD5, we now check the size() which has to be 16.
Fixed the destruction of a revision entry, the destroy event would not properly eliminate the reference to the files table if there was one and delete the file if there was no more references to it.
Started removing a few controlled_vars in the (super slow) fade out of that library. C++11 gives us inline initialization of variable members which is better (except for the fact that it is not mandatory, but Coverity can help with that).
Updated the feed and snap software description plugins so it makes use of the new auto-delete old revision feature.
Added "const" in various places.
Various clean ups.

SNAP-168: Fixed the loop in snap_backend::process_timeout() which would skip 4 websites on each iteration! This happens when the lock() call fails.

Fixed the name of the lock in snap_backend::process_backend_uri() so each action has its own specific lock.
Fixed a bug in the handling of the cached messages: they would never get deleted.