SNAP-310: Moved the locale default settings to the locale plugin instead of the locale_settings plugin so that way we get a default locale from the first installation.
Changed the dependencies of the locale plugin to include the content plugin.
Added access to the f_ready flag of the snap_child object so we know whether we are inside the path::execute() or initializing the database.
Call the finish_update() in the layout.cpp only if snap_child is ready.
The links and content "modified" signals are not running only if the snap_child is ready. Otherwise they are completely ignored.
Fixed the output and handling of the page status in dbutils.
Removed a duplicate of the website/info permission definition in info/content.xml.
Fixed the ownership of the "/images/users" to "users" in the users_ui/content.xml file.
A few other cleanups too.