SNAP-291: Fixed the fact that some plugins may be "installed" via the "default_plugins" list in snapserver.conf (we could not install/remove anything in that state.)
Added a checkmark to clearly show that a plugin is installed, although there is a title attribute.
Prevent all Install and Remove buttons when the snapserver.conf makes use of an authoritative "plugins" list.
Fixed the error() function of snap.cgi so the error actually shows up on the screen and not just in our logs.
Various clean ups.