SNAP-316: Added a "sortable" widget to allow for a list of items to be sortable by dragging and dropping items in a list.
The sortable includes a 3rd party MIT licensed JavaScript doing all the low level work that functions under all modern browsers.
Simplified a few editor entries referencing the name of the item for an attribute since we can use the {$name} which is smaller and probably faster.
Fixed the editor wasModified() test because some widget originalValue_ needs to be tested and not their HTML (that includes the new sortable widget).
Changed some typeof value to check for undefined as just plain 'value !== undefined'.
Changed the event on all widget types that requests a checkModified() call, we now make sure to create only a single timer.
Fixed the password exception name from "versions" to "password".