Fixed issue with CMAKE_MODULE_PATH
With CMAKE_PREFIX_PATH, we don't need to pass in the module path from
the command line--what we can do is figure it out based on the prefix.

So all modules are installed under:
"dist/share/cmake/<blah>/<blah>Config.cmake"
So all we need to do is make sure CMAKE_PREFIX_PATH is set to the "dist"
path.

I also discovered that zipios2 and as2js slipped through the cracks
cmake-module wise. So that has been fixed as well.