Fixed bug in the libQtCassandra library, and added new functionality in snapdb.
When calling "dropContext()" on a context, the cassandra pointer was
being released before clearing out the cache. This has been fixed.

I added two new functions to snapdb: "dump_context()" and "restore_context()."
This is in relation to SNAP-60. The functions are not complete at all,
but there is a starting point for "dump_context()". Right now it dumps
out an XML file of all of the tables in the snap_websites context.