EX-56: Enhanced the compressor handling so we can get any compressor.
Implemented a "deflate" decompress() function, had to add a function that accepts the uncompressed_size parameter.
Optimized the BEST_COMPRESSION implementation.
Fixed the decompress() function which would not set the compressor_name variable properly in all cases.
Removed the #include of poison.h from two headers, it cannot be used in headers, only in .cpp files.
Fixed the compression_expcetion name to compression instead of snap_child.
Added a fix to the force_locale() function: clear the revision key so it gets regenerated if the locale gets changed.
Various cleanups.