Now --dump-context works.
When using the --dump-context function, the output of the entire context
is streamed to either the stdout, or to a file if specified.

Also added a small change to the table_model class which puts the
default rows on the contructor, but allows the user to change them on
construction of the model itself.