SNAP-110: Removed the q=... parameter when running an installation. The path is actually enough and q=... is ignored now.
Added support for RFC-3339 date and time in our websites. (the dcterms.date and Co. support the time too!)
Updated our headers to make use of the new long date where possible.
Added the email address and identifier of the author in the snap header data.

SNAP-150: Got the RSS and Atom feeds to both validate with http://feedvalidator.org/check.cgi and get approval even if a few things are missing.
The RFC-3339 date and time is reused in our feeds to send a full date and time instead of just the year/month/day.
Applied a fix for xml:lang which does not work right in the Qt XSLT 2.0.
Fixed the base=... for atoms, it would use the last page URL instead of the item URL.
Added the author block because we do not validate without it. The user name and email are bogus when undefined on a page.
Fixed the date3339 used as the creation date of an Atom to include the time and timezone.
Fixed the path to the feed default logo so it is a full path instead of being relative.