SNAP-259: Added the code for the password policy check on the client side.
Fixed the "special" entry as it was expected to be plural.
Properly reset the counters just before counting the various types of characters in a password. The policy_t class would never reset those counters.
Added the missing test of the global minimum number of characters in a password.
Also added that field in the settings were it was not defined at all.
Swapped a couple of fields in the settings as it seems to make more sense that way (At least to me...)
Added the list of missing fields for the tab key to work properly in the password settings.
Added a standalone project in the output plugin to generate the javascript-unicode.js code (it is about 17Kb of data about Unicode character categories.)
Various clean ups.