Fixed error where nothing was being dumped.
Reversed the test, so now rows are dumped. I encountered an error with
the files table entries, so I employed snap::dbutils. This uncovered a
crash, so I added a feature where just a specified number of tables on
the command line can be specified (--tables-to-dump). The code is
untested, but compiles.