SNAP-163: Started work to add a Year+Month date widget.
Moved the Date Widget WYSIWYG test to the date_widgets plugin (since it is not always available in the editor plugin.)
Added a dropdown widget (for the Year+Month), although it is not yet tested.
Renamde the WYSIWYG Editor page as just Editor Settings (to better match other plugins).
Added logos to the "flash" and "mailinglist" plugins. Also created a logo for our future webform plugin.
Properly allow attributes on the item element in simplied XSLT 2.0 DTD.
Applied a few clean ups to the mailinglist plugin (wrong name, install the content.xml, various comments).