SNAP-163: Added a dropdown date edit widget to the date widget test page.
Added support for an algorithm to place the dropdown.
Added support for a "transparent" class so composed widgets do not get the wrong colors.
Added CSS so the regular dropdown and dropdown date widget is properly formatted.
Added JS to handle the dropdown date edit widgets (not finised, we do not yet properly collect the date though.