SNAP-163: Added support for dropdowns using a scrollbar once too many items are present in it.
Fixed bug with new_class definition that searched for the first '-' instead of the '='.
Limit max_count to the number of items in all the column based algorithms.
Moved the column-count to the ul tag instead of the wrapping div.
Fixed several outerHeight() which were actually expected to be outerWidth().
Fixed up the cleaning up of the dropdown div/ul/li elements as some things were missing/changed.