SNAP-163: Finished up the dropdown date widget.
Added a new editor data type "dropdown-date" which supports a date as "YYYY/MM/DD" as that widget returns.
Fixed the date-form.xsl so reloading a dropdown date works as expected.
Actually implemented the JavaScript code for the dropdown date.
Added new fields used by the editor to dbutils.cpp.
Changed the editor algorithm so the date validation first tries to convert the value, if that retruns a value of type "date", then we use that value directly.