Took out TTL and WRITETIME functions. Added count feature.
The TTL() and WRITETIME() CQL functions never worked right. So I took
them out, and it looks like they are not needed at all. Alexis stores
those values in the "session" table.

Added the "--count" for the SELECT from tables, and added it to the
LIMIT clause. This allows the user to control how many records are
pulled for the dump. Just specify a huge number, like 10,000,000 and you
should get everything for the context dump.