SNAP-163: Started adding a compiling e-Payment credit card plugin.
Fixed the dropdown algorithm for the dropdown date edit widget.
Fixed the dropdown date edit, the year, month, or day widgets may not be defined and the JavaScript has to take that in account.
Fixed the getLabel() so the ':' gets removed if present in the label so they do not appear in a messages.
Fixed the messages copy operator where the name of the widget was not being copied and thus some labels would not make it in the client's side.
Fixed the autoLogout_() function which would not actually use the new URI with the hit=transparent parameter.
Enhanced the preventAutoLogout() to test whether the auto_logout_id_ parameter is already a NaN.