SNAP-327: Fixed the various problems with the list not waking up "more often".
The add_uri_for_processing() function expects the 'action' parameter to be set to the action including the namespace (i.e. "list::pagelist" instead of just "pagelist").
Reverted the setCount(100) in process_timeout() back to setCount(1). The inner loop was NOT taking the possibility to receive more than one cell at a time and it would skip some rows.
Fixed the calulation of date_limit in on_backend_action() so it does not become 0 or close to 0. We were expected to use 'f_date_limit' itself, not 'f_date_limit - current-date'.
Added a 1 minute time limit to process the first 100 items found in the list table. Once processing reaches 1 minute, the process stops, whether the 100 items were worked on or not.
Rephrase the information about the algorithm so it is clearer and includes the new additional timeout scheme.