SNAP-168: Applied a fix to the snap_child::udp_ping() to send the website URI and not the current page full URI with path, query string, anchor...
Added a get_website_uri() to the snap_uri class so we can get the website URI directly, optionally with the port indicated if different from the default according to the protocol.
Removed the sendmail PING name since we now use a function which ultimately take care of that somewhere else.