EX-92: Added the Tax ID document upload with a new dropped-file widget.
Added the "dropped-file" widget type to the editor. (widget, CSS, JavaScript...)
Added such a widget to the editor test page.
Actually implemented a way to download any attachment as such (instead of directly showing in the browser) by using ?download=attachment in the query string area.
Implemented the MIME type to icon and path in the editor plugin using a new function in the mimetype plugin.
Enhanced the Drag and Drop so the drop works anywhere in the widget, not just the editor-content area (very useful for such widget). This includes a dragCounter_ which ensures proper dragenter and dragleave functionality.
Fixed a few regular expression and simplified others.
Finally moved the filter_filename() "inner-function" to the filter plugin. It is now used in two places.
Fixed the height of the attachment icon box used in dropped file with preview.
Removed the "type" parameter from the attachment tag in the editor widget definition. You are expected to use the identification attribute instead.
Fixed the snap.cgi handling of "\n", we MUST send all of them or we could end up removing 0x0A bytes from binary files being POSTed.
Fixed the default path to the snap.cgi in our apache2.conf sample.
Fixed the change from any save mode to AUTO_DRAFT when we are saving an attachment, because that's not correct. I added an AUTO_ATTACHMENT although it is not yet implemented.
Fixed the handling of attachment saving in the editor, it was not taking the identification and force-filename/path parameters in account.
Fixed a bug in editor.js where I completely misused substr().
Added all the missing variables to the openDropdown() function.
Various clean ups.