Work in progress. Now removing the QCassandraPrivate object.
I know I brought it back, but I learned a few things about how it
populates the cassandra_types objects and uses them for table meta-data.
So I'm moving the guts of that object into the super-structure of the
class objects to make handling the legacy-way of meta data easier in a
backwards-compatible way.