SNAP-332: Partial saving was not duplicating the Revision data; this happened when a page with branch/revision 0.0 was saved to branch/revision 1.0 because the revision is the same.
Fixed the action to "administer" for any path under /admin even if the path action was defined (except for layout pages). This may be a problem in some situations, but it was a potential security issue.
Fixed the move_page() function which would marked the old page as DELETED instead of MOVED. That way the redirect will work as expected.
Fixed the path plugin to read the CLONE link and not ORIGINAL PAGE, it was swapped.
Fixed the display of temporary form data saved whenever the form included some invalid fields. These are not transformed so all the data are strings.
Reduced the number of columns read in the copy_row() to only 100 to make sure we do not use too much RAM.
Fixed/add some names to the CT_int64_value.
Added a moved-from attribute to define a way to "move" an administrative settings page (it does not move the old page, but marks it as moved and create a new version--the old page will redirect the user).
Various clean ups.