SNAP-199: Transformed the not_modified() function to support If-None-Match AND If-Modified-Since.
Removed the ETag support for JS and CSS files. That was "wrong". We just need a 1 year duration (max-age=1 year in seconds).
Added the ETag support with other attachments, along a small (1h for now, should be a parameter) cache duration.
Removed the Content-Type from 304 replies where it should never be defined (as per HTTP/1.1).
Removed the "X-Content-Type-Options: nosniff" from the 304 since I do not think it is supposed to be there.
Added the "X-Content-Type-Options: nosniff" to regular 200 replies with JS and CSS files.
Properly exit with 0 when sending a 304 reply (we would exit(1) instead!)
Added a check of the day against last-day-of-month, if too large the string_to_date() returns false instead of an adjusted date.
Noticed that the YYYY-MM-DD test to convert an HTTP date to time_t should be checked first because it does not make sense after the week day check.
Fixed all the -1 in the cache_control_settings to use IGNORE_VALUE instead.
Fixed the documentation of the cache_control_settings which still had many copy/paste errors and was a bit weak in many places.
Removed the ability of using -1 to set AGE_MAXIMUM. Instead, you now have to use AGE_MAXIMUM (duh!)
Removed the check on client no-cache/no-store within the no_caching() function, because that's for proxies and not the client himself.
Moved the 'no-transform' so it gets added with caching happens or not (although it is expected to be for proxy caches, you never know what can happen in between.)
Fixed test of md5 size using size() == 16 instead of !nullValue().
Updated the silktide .js file with the newer version which made it work again as expected.
Tried to add a "let's clear permissions on a client Reload, but it is so slow that for now I commented out the one line that makes it happen. It works as expected otherwise.
Various clean ups.