Fixed found bug in simple_counter and context_management.
First of all, when creating a context, the entire schema must also be
created. So tables are created as well, if they are defined properly in
the class/object hierarchy.

Found a big in QCassandraCell where the wrong value was being written to
the database. This resulted in bad results when the cache is cleared and
results reread.