EX-11: Added editor support to round dates up or down so partial dates can be checked properly (with a one month precision as for credit cards.)
Fixed the + X month and + X year in the date parser to make use of actual month and year (tm_mon/tm_year) and not just an approximation.
Fixed the date parameter which was a plain integers set to 0, 1, 2, or 3 with an enumeration.
Removed the error when a regex is given a name that Core (editor.cpp) does not know about. That allows for other plugins to implement that feature.
"Optimized" a loop by running it only if required.
Removed the two large blocks of code that were used to parse dates and time in the editor.cpp. We now use the trusted locale parse_date() and parse_time() functions which have been in place for a while and have not generated any problems.
Marked alll messages->set_error() calls with "is_secret" instead of "false" when validating data (the message may include the contents of a secret field!)
Fixed one set_error() which would use the value of a field as is instead of having it passed through html_64max().
Optimized a JavaScript by saving a list of items in a variable instead of gathering that list each time I run a loop.
Fixed the width check of dropdown with sliderbars, the +20 pixels has to be applied there too (still need a way to measure that scrollbar though.)
Fixed the dropdown opening at the bottom by not forcing the height if the auto height is smaller than what is available below the dropdown.