SNAP-163: Worked some more on the credit card form so it works as expected.
Fixed all the editor forms with a proper timeout, and for the few that need it, a proper auto-reset.
Changed the timeout to 1 day for most forms. 1 hour for more important data (such as the anti-hamering form).
Added new epayment int8 fields to the dbutils.
Fixed the session retrieval it should have used a value-of instead of copy-of.
Added the "composite" class to mark composite (shell/encapsulating) widgets as such, which is very important when doing a restore.
Fixed the date widget that was declaring two tags withe class 'editor-content' and put the 'value' attribute in the wrong one.
Fixed the editor restoreValue() function, we really cannot use the setValue() at all.
When restoring all the values, we now properly hide the Save Dialog.
Fixed the auto-reset of forms: (1) it would be triggered only once; (2) it would not be stopped when the form timed out; (3) it would not mark the form as unmodified.
Added missing declaration to the epayment credit card CMakeLists.txt (validate XML and XSL files and added them as dependencies.)
Added epayment credit card settings. We can hide 3 fields and define a default country name.
Prevent the auto-save of the credit card form (even though it would not save any of the credit card data, it would create a new revision each time.)