SNAP-163: Got the epayment_stripe plugin to compile. It shows up as a payment gateway in the credit card settings page.
Changed the save_editor_fields() signal to make use of a structure so we can pass many parameters without making it super complicated. We also get the error flag (i.e. if validation failed has_errors() returns true.)
Added a Phone field to the credit card form (necessary if shipping with FedEx, etc.)
Added a hidden field to the credit card form with the name of the gateway (the plugin, really.)
Put the user-messages in the bare and white themes so we see error messages when they happen.
Fixed a bug in the editor form implementation, there was one case where /node() was missing so the item tag of a dropdown would appear in the HTML which was wrong.
Also fixed a potential bug with the modified flag while saving an editor form: a plugin could make further changes and the editor would not detect such before.
Found a bug in the editor.js where I would use 'e.target' instead of 'this' and would end up with the wrong tag as the selected item of a dropdown widget.
Enhanced the link raised exception when the link_back is wrong to include the source and destination keys.
Added a safe_timezone class so we can change the timezone and it gets restored even if an exception occurs.