SNAP-163: Finished up the form for Credit Card data.
Added client side code and CSS to properly handle the credit card form in the popup.
Added a public instance of the credit card form under /epayment/credit-card.
Updated the settings to include some additional options (i.e. you may want a single address instead of a business + delivery addresses).
Added a business name, "attention" field, a billing and delivery address instead of just one address, and a phone number.
Fixed the epayment buttons look and feel (CSS), order, and actually implemented them in the Credit Card popup window.
Finally fixed the darkenPage() using a proper stack so we can now have any number of popups and the darken page (renamed screen) works as expected (i.e. gets shown, moves in z-index , gets hidden.)
Updated the server_access plugin to make use of the new darkenScreen feature internally. This makes life much easier and things are way cleaner this way (i.e. the new stack of darken pages allows us to not interfer between the wait darken screen of the server access and the darken screens of popups.)
Updated the editor plugin to make use of the new darken screen implementation.
Fixed a bug in the editor.restoreValue() function, the "value" attribute has to be checked against "" and not just as is.
Added the form owner and id to the editor XML data so one can check them in their XSLT code and add them to the form so a client side JavaScript code knows which form it is dealing with.
Cleaned up definitions of the ResultData used in the server access callbacks.
Added the ePaymentInstance variable declaration (it was missing.)
Fixed the epayment facility items, (1) moved the size to the CSS file so it can easily be overridden, (2) put the name of the facility in each button, and (3) added a 'clear both float' tag at the end.
Removed the on_process_post() from the stripe implementation since for credit card processors we do not use such (we may re-add it later if we implement bitcoins through them.)
Cleaned up the PopupData definition, especially, put it in one single place!
Fixed the closing of a popup from within the popup IFRAME without having to know the popup identifier. This will simplify many popups in the long term!
Added a few "administer" permissions so we can submit the Credit Card form. It should use the "submit" action instead of "administer", though.
Applied a few fixes to 3 other permissions which I found were wrong (wrong name for the link back and one was a duplicate page definition.)
Various clean ups.