SNAP-163: Got the direct (although, with customer so we can create a subscription) payment with Stripe working.
Properly redirect the user to the Thank You page once the Stripe payment succeeded.
Adjusted the maximum number of digits in a CVV as per Stripe: 3 or 4 instead of 2 to 7.
Fixed the Stripe settings, we actually get a test key when registering.
Fixed the credit card form: some fields should not have been required, added filter to country, fixed some language.
Added a Luhn check to the credit card number before sending it to the server. That way we avoid many mistakes from the very source.
Added the source URI to the credit card query strings (from=<uri>) so that way we have the URI to the invoice.
Added a page of settings to the e-Payment plugin. We can now prevent all gateways from getting tokens to re-charge the card. Avoiding that liability.
Added a "country filter" to the editor so we can make sure a field is the valid (English) name of a country.
Allow for a redirect on a form timeout. This is important for administrative forms where the data is still visible after a reset.
Put an e-payment grand total in the invoice so we are not forced to check out all the product and redo the math in each processor.
Fixed the Paypal checkout, if the invoice number comes out as 0, there is no invoice so we have to not initiate the payment.
Fixed the ajax_redirect() call so if the redirect_url is empty, we do not attempt to do a redirect.
Various clean ups.