Now implements consistency and timestamp.
Added features to the QCassandraQuery object which allow you to set the
consistency and the timestamp for the current active query.

Also, put back the code I took out regarding consistency for the
QCassandraLock class. Likewise, the support for consistency and
timestamp for tables, rows and values has been restored and now function
against CQL.