EX-95: Put back a setConsistency() that was removed by Doug.
Fixed the default consistency level, it has to be at least QUORUM (would work with ALL), but DEFAULT is often ONE so it is wrong.
Put the documentation on the consistency level back to the QCassandraLock.
Commented out the std::cout in QCassandraQuery to avoid all the output... Let's try not too commit it.