EX-95: Bumped libQtCassandra version to 0.6.0 to clearly show we are using a version with CQL.
Fixed a bug in the clearCache() of the QCassandraRow implementation, it also needs to clear its own f_query object (like the Table does).
Just in case, replaced all toStdString() with toUtf8().data(). There was at least one case that needed it.
Use the join(",") to create the list of hosts.
Various small clean ups.