EX-95: Fixed several problems with the new CQL implementation of libQtCassandra.
Fixed the readRows() and readCells() functions, both have to clear the cache of cells before getting the following page, but not clear the f_query pointer.
Fixed the QCassandraTable::exists() function, it has to protect the existing f_query and use a new one for its own purpose.
Added the consistency and timestamp to the query in QCassandraTable::getValue(). The timestamp may not be required, but the consistency is, for sure.
Fixed the predicate so we can have the start and/or end set to empty. (It is often used that way in Snap, probably because of the way it was implemented in thrift... i.e. you had to have a key or a range.)
Added a test against value in getByteArrayFromValue() so if nullptr we cleanly throw.
Added a getByteArrayColumn() with a const char * to avoid the conversion to QString and back. (At this point others are not used so the point is moot in fixing those.)
Various clean ups.