iEX-95: Remove usage of QByteArray::fromRawBuffer() since that function is rather unsafe.
Added support for the f_reversed predicate so looking for JavaScript and CSS files works as expected (i.e. we get the latest first.)
Kept some code testing various metrics to slow down the INSERTs but that did not help, the database died non-the-less... so it is still there but all commented out.
Various clean ups.