EX-95: Going to CQL, reseting the entire database, I noticed that mimetype was missing in the list of minimal core plugins to be installed.
Added code to prevent adding a page if its parent does not exist in the on_save_content() function (i.e. your content.xml must be correct in that regard.)
Inverted the start and end search keys because when we use the reversed flag in CQL we do not swap the start/end valus in the WHERE clause.
Fixed the dependency declaration of test-plugin.js. It had that on one line instead of separate tags.
Fixed the fact that we do not want to delete the "lock_table" with just --drop-tables in snapdb.
Fixed the fact that if some tables did not make it right, attempting to drop them creates an exception so we need to catch those and go on with deleting the other tables.