EX-95: Fixed the memtable_flush_period_in_ms from minutes to milliseconds.
Removed the Timestamp, which cause some slowness and we really only use it on the DELETE (dropRow and dropCell calls.)
The sendmail plugin would not properly initialize its table. We now have an initial_update() doing so.
Create the backend table along the domains and websites in the server (snapwebsites.cpp), that way when we start the backend processes the table already exists.
Various clean ups.