EX-95: Fixed the memtable_flush_period_in_ms from minutes to milliseconds when creating the lock_table.
Removed the Timestamp, which cause some slowness.
Changed the DELETE to always use the most current time when dropping rows and cells.
Added a test to make sure that the gettimeofday() function works.
Added a QCassandraTimeout class so we can change the timeout of a query in RAII.
Increased the timeout to 5 minutes when doing a CREATE TABLE to "make sure" we get the table each time.
Various clean ups.