SNAP-215: First version of an Cassandra Proxy implementation.
The library now includes 3 new classes: Proxy, Order, and OrderResult. The Proxy manages orders and results sent between clients and the proxy daemon.
Converted the QCassandra, QCassandraContext, QCassandraTable, and QCassandraRow to use the proxy server only.
Updated the QCassandraPredicate to handle the blobification of predicates and send them (as CQL) to the proxy.
Changed the retrieval of contexts as it would retrieve the data twice; also we still need to look into a way to completely avoid the DESCRIBE CLUSTER function...
Completely removed the 'snitch', it was only for display and thus far I had no other use for that parameter.
Removed the "version" from the SessionMeta since it was not used (less to travel between client/server).
Added encode/decode functions to help with sending queries from the QCassandra objects to the snapdbproxy and send replies.
Removed the COUNTER capability. We do not use it and it is not really sensible anyway.
Fixed the getStringColumn() functions which would return ASCII instead of UTF-8 as expected.
Various clean ups such as documentation updates.