NAP-215: Created the snapdbproxy so we can send CQL commands to any Cassandra Node without having to connect to more than one TCP/IP connection.
Added snapdbproxy.conf/properties and updated the logrotate settings to handle the new files.
Added snaplock.conf/properties and updated the logrotate settings to handle the new files.
Added the server_name=... sample to the snapinit.conf file, the only place where it can be defined.
Removed the server_name from the snapserver.conf, it is now forbidden to have that there because we get the info from snapinit.conf via --server-name.
Added documentation for the <safe> tag in the snapinit.xml file.
Added the snapdbproxy and snaplock tools to the snapinit.xml file. Both are now required and need to be started before we move forward.
Added a kill() to the snap_thread implementation so we can send a signal to a thread (assuming that the signal will somehow be handled as expected...)
Cleaned the add_subdirectory() for snapdb and added snapdbproxy and snaplock.
Removed/cleaned up various comments about exceptions, especially the one talking about thrift.
Created a fatal_error() function in snapinit.cpp so we can get the error (1) in our logs, (2) in syslog(), (3) on the console, as required.
For developers, added support for daemons that appear one directory removed (i.e. .../snapdbproxy/snapdbproxy instead of just .../snapdbproxy).
snapinit now sends a DIED message whenever a service dies and snapcommunicator is still running.
snapinit now uses a switch() to process its messages faster.
Updated the snaplayout code connecting to Cassandra, it now connects to snapdbproxy.
Fixed the default port in snapdb to connect to snapdbproxy. We may change that back to Cassandra if we do a direct version. TBD.
Added a test to verify that the new lock implementation: i.e. snaplock, works as expected. (the test is the one we had with libQtCassandra revamped to work with the new interface.)

SNAP-168: Heavily updated the snap_communicator environment (library and daemon) to be able to handle the lock and snapdbproxy.

We now support a way to reply to the sender (i.e. A -> SC -> B, B -> SC -> A).
Allow for the server to be indicated as the destination, so we can specifically send to a named server instead of any server that has at least one instance of a service.
Properly ignore messages to services that are local but not currently registered.
Now a TCP client buffer socket can be kept blocking so that way we do not have to have a complex run() loop for the locks.
Fixed the priority type, it should have used priority_t and not just int.

SNAP-341: Fixed the place where the setup_firewall() function gets called and moved the SAFE message after that too.

Cleaned up the firewall implementation as I was updating the code to support the snapdbproxy service.
Especially, it is not a singleton so no need for a create() function.
And moved various global variables that could just be variable members to the class.

Various clean ups and documentation enhancements.